+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Downgrade php version to support php 7.2 to 7.4
+
 ### Fixed
 
 - Fix resolve refunded payment and return view