+++ CHANGELOG.md
@@ -6,12 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-### Changed
+## [1.9.0] - 2023-03-07
 
+### Changed
 - Downgrade php version to support php 7.2 to 7.4
 
 ### Fixed
-
 - Fix resolve refunded payment and return view
 
 ### Added